# MultiplayerSynchronizer: Improve performance of watched variables