--- a/modules/multiplayer/scene_replication_interface.cpp
+++ b/modules/multiplayer/scene_replication_interface.cpp
@@ -709,16 +709,18 @@ void SceneReplicationInterface::_send_delta(int p_peer, const HashSet<ObjectID> 
 	for (const ObjectID &oid : p_synchronizers) {
 		MultiplayerSynchronizer *sync = get_id_as<MultiplayerSynchronizer>(oid);
 		ERR_CONTINUE(!sync || !sync->get_replication_config().is_valid() || !sync->is_multiplayer_authority());
-		uint32_t net_id;
-		if (!_verify_synchronizer(p_peer, sync, net_id)) {
-			continue;
-		}
+
 		uint64_t last_usec = p_last_watch_usecs.has(oid) ? p_last_watch_usecs[oid] : 0;
 		uint64_t indexes;
 		List<Variant> delta = sync->get_delta_state(p_usec, last_usec, indexes);
 
 		if (!delta.size()) {
 			continue; // Nothing to update.
+		}
+
+		uint32_t net_id;
+		if (!_verify_synchronizer(p_peer, sync, net_id)) {
+			continue;
 		}
 
 		Vector<const Variant *> varp;