@DennisManaa 2023-09-13
It's also important to emphasize that the performance enhancement in `Main::iteration` is not simply a 5% improvement. To accurately gauge the true performance gain, it's crucial to shift our focus to the unchanged portion in our comparisons rather than the altered one.

Assuming that the number of required cycles for the code outside of `Main::iteration` remained relatively constant, we can establish the following relationship:

```math
(1 - 0.748) \cdot \text{[old\_total\_workload]} = (1 - 0.691) \cdot \text{[new\_total\_workload]}
```

For the sake of simplicity, let's assume the old total workload was 1, leading to the following equation:

```math
\text{[new\_total\_workload]} = \frac{1 - 0.748}{1 - 0.691} = \frac{0.252}{0.309} \approx 0.815533981
```

This reveals that the total workload attributable to `Main::iteration` in our application has actually decreased by approximately 18%, a far greater improvement than the initial perception of just 5%. This conclusion is further substantiated by the observations regarding the "aggregated sample costs" of 1.674E+11 (before) and 1.294E+11 (after), as demonstrated in the provided screenshots.

However, it's important to note that the assumption that the code outside of `Main::iteration` remained completely constant doesn't hold entirely true. It's nearly true, but not entirely accurate.

### Before
![Before](https://github.com/godotengine/godot/assets/26871131/82852622-d06b-49b6-9fa0-8edf51f5854a)

### After
![After](https://github.com/godotengine/godot/assets/26871131/93f75aa3-5027-45a2-9d7c-7f43b186100d)