@Faless 2023-10-04
@DennisManaa thank you very much for the comprehensive report.

As I mentioned in chat, this PR does indeed introduce the side effect of skipping the first (few) delta synchronizations, resulting in an out of sync state until those variable changes.

I've opened https://github.com/godotengine/godot/pull/82777 which should address `_verify_synchronizer` being so slow among other things (the function itself is removed, the actual optimizations are in `send_object_cache`, `make_object_cache`, and `is_cache_confirmed`).

Would be great if you could give that a try and confirm that it does work for you, and that you do get the expected performance improvements.